backend/requirements.txt
blinker==1.8.2
click==8.1.7
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
nest-asyncio==1.6.0
pyaes==1.6.1
pyasn1==0.6.0
python-dotenv==1.0.1
rsa==4.9
Telethon==1.35.0
Werkzeug==3.0.3
